+++ CHANGELOG.md
@@ -7,7 +7,13 @@
 
 
 # Next
+
+
+
+# 3.2.0
+> 2020.07.21  
 1. <font color=orange>修改</font> `noCache`默认值改为false，阻止IE缓存改为添加请求头`Cache-Control:no-cache`,`Pragma:no-cache`，并且`noCache`参数将会在下个版本`4.0.0`废弃。
+2. <font color=orange>修改</font> `processResponse`在http状态码`204`时也执行。
 
 
 # 3.1.2